+++ mj-calc/src/App.tsx
@@ -1,12 +1,10 @@
-import './App.css';
-import './index.css';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import Calculator from './modules/Calculator';
+import "./App.css";
+import "./index.css";
+import "bootstrap/dist/css/bootstrap.min.css";
+import Calculator from "./modules/Calculator";
 
 function App() {
-  return (
-   <Calculator />
-  );
+  return <Calculator />;
 }
 
 export default App;
